refactor(analysis): split effort gitattributes owner
